Implemented right click class list interval selection
Previously, multiline selection was removed for being confusing since only one class bytecode was loaded anyways.
Also previously, the ability to copy class names (and subsequently class info) was introduced.
This commit allows to select multiple classes, but only by holding RMB and only to bring up the Copy Popup menu.
The selection then resets to what was selected before, which should correspond to the bytecode loaded in the buffers.